requirement.txt
asgiref==3.2.7
Django==3.0.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.1
